error catch on user null input to AdminView foodname

+++ Ecomerce_Restaurant/Ecomerce_Restaurant/Models/FoodModels/FoodCategories.cs
@@ -45,7 +45,7 @@ namespace Ecomerce_Restaurant.Models.FoodModels
         
         
         [Required]
-        public double FoodPrice { get; set; }
+        public decimal FoodPrice { get; set; }
         
         
         [Display( Name= "Current Rating")]
